requirements.txt
Flask==3.0.0
openai>=1.30.0
python-dotenv==1.0.0
gunicorn==21.2.0
